chore: forge fmt

+++ testdata/cheats/Json.t.sol
@@ -105,7 +105,7 @@ contract ParseJson is DSTest {
     }
 
     function parseJsonUint(string memory json, string memory path) public returns (uint256) {
-        uint data = vm.parseJsonUint(json, path);
+        uint256 data = vm.parseJsonUint(json, path);
     }
 
     function test_coercionUint() public {
